requirements.txt
langchain-community 
langchain-openai
chromadb
pydantic
pandas
fastapi[standard]
duckduckgo-search
wikipedia